AI_server/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
transformers==4.44.2
torch==2.4.0
Pillow==10.4.0
numpy==2.1.1
huggingface-hub==0.24.6
httpx==0.27.2
